requirements.txt
flask
flask-cors
flask-socketio
eventlet
mysql-connector-python
python-dotenv
bcrypt
sqlalchemy
PyJWT
tenacity
google-generativeai
google-auth